+++ acf/php/mod-manualinput-settings.php
@@ -27,6 +27,26 @@ if (function_exists('acf_add_local_field_group')) {
             'ui_on_text' => '',
             'ui_off_text' => '',
         ),
+        1 => array(
+            'key' => 'field_68c91a16001ab',
+            'label' => __('Visa som slider', 'municipio-customisation'),
+            'name' => 'show_as_slider',
+            'aria-label' => '',
+            'type' => 'true_false',
+            'instructions' => __('Visar korten i en slider. Kolumninställningen styr hur många slides som syns. Fungerar med Kort, Block, Box, Segment och Nyheter.', 'municipio-customisation'),
+            'required' => 0,
+            'conditional_logic' => 0,
+            'wrapper' => array(
+                'width' => '',
+                'class' => '',
+                'id' => '',
+            ),
+            'message' => __('Yes', 'municipio-customisation'),
+            'default_value' => 0,
+            'ui' => 0,
+            'ui_on_text' => '',
+            'ui_off_text' => '',
+        ),
     ),
     'location' => array(
         0 => array(